requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
feedparser==6.0.11
python-dotenv==1.0.1
pillow==10.3.0
